Fix surface rendering problem
There was an error to normalize vertex
Unintended conversion from float to integer

+++ src/datavisualization/axis/qvalue3daxis.cpp
@@ -270,7 +270,7 @@ float QValue3DAxis::labelPositionAt(int index)
     return formatter()->labelPositions().at(index);
 }
 
-float QValue3DAxis::positionAt(int x)
+float QValue3DAxis::positionAt(float x)
 {
     return formatter()->positionAt(x);
 }
